+++ css/_profile.scss
@@ -7,3 +7,11 @@
 .custom-control {
   margin-top: 30px;
 }
+
+.custom-control-label {
+  cursor: pointer;
+}
+
+.download {
+  width: 400px;
+}
