+++ css/base.css
@@ -431,6 +431,12 @@ nav {
 .custom-control {
   margin-top: 30px; }
 
+.custom-control-label {
+  cursor: pointer; }
+
+.download {
+  width: 400px; }
+
 .game-page .game-description .row {
   position: relative;
   top: auto;
